// packages/simmerbase/src/schemas/species.ts
// Auto-generated by create-types-files.ts — 2026-02-13T00:41:02.350Z
// Do not edit manually.

import z from 'zod';

export const SpeciesBaseRowSchema = z.object({
	description: z.string().nullable(),
	genus_id: z.uuid().nullable(),
	id: z.uuid(),
	species_name: z.string(),
});

export type SpeciesBaseRow = z.infer<typeof SpeciesBaseRowSchema>;

export const SpeciesInsertSchema = SpeciesBaseRowSchema;

export type SpeciesInsert = z.infer<typeof SpeciesInsertSchema>;

export const SpeciesUpdateSchema = SpeciesInsertSchema.omit({ id: true }).partial();

export type SpeciesUpdate = z.infer<typeof SpeciesUpdateSchema>;