// packages/simmerbase/src/schemas/tag_groups.ts
// Auto-generated by create-types-files.ts — 2026-02-13T00:41:02.350Z
// Do not edit manually.

import z from 'zod';

export const TagGroupsBaseRowSchema = z.object({
	id: z.uuid(),
	tag_group_name: z.string(),
});

export type TagGroupsBaseRow = z.infer<typeof TagGroupsBaseRowSchema>;

export const TagGroupsInsertSchema = TagGroupsBaseRowSchema;

export type TagGroupsInsert = z.infer<typeof TagGroupsInsertSchema>;

export const TagGroupsUpdateSchema = TagGroupsInsertSchema.omit({ id: true }).partial();

export type TagGroupsUpdate = z.infer<typeof TagGroupsUpdateSchema>;